Improve tests and logging (#682)
* Add retry decorator
* Split tests
* Improve readability in ci/cd

+++ tests/test_keras_integration.py
@@ -14,14 +14,17 @@ from huggingface_hub.keras_mixin import (
     push_to_hub_keras,
     save_pretrained_keras,
 )
+from huggingface_hub.utils import logging
 
 from .testing_constants import ENDPOINT_STAGING, PASS, USER
-from .testing_utils import set_write_permission_and_retry
+from .testing_utils import retry_endpoint, set_write_permission_and_retry
 
 
 def repo_name(id=uuid.uuid4().hex[:6]):
     return "keras-repo-{0}-{1}".format(id, int(time.time() * 10e3))
 
+
+logger = logging.get_logger(__name__)
 
 WORKING_REPO_SUBDIR = f"fixtures/working_repo_{__name__.split('.')[-1]}"
 WORKING_REPO_DIR = os.path.join(
@@ -65,10 +68,11 @@ else:
 @require_tf
 class HubMixingTestKeras(unittest.TestCase):
     def tearDown(self) -> None:
-        try:
+        if os.path.exists(WORKING_REPO_DIR):
             shutil.rmtree(WORKING_REPO_DIR, onerror=set_write_permission_and_retry)
-        except FileNotFoundError:
-            pass
+        logger.info(
+            f"Does {WORKING_REPO_DIR} exist: {os.path.exists(WORKING_REPO_DIR)}"
+        )
 
     @classmethod
     def setUpClass(cls):
@@ -140,6 +144,7 @@ class HubMixingTestKeras(unittest.TestCase):
         model = DummyModel.from_pretrained(f"{WORKING_REPO_DIR}/{REPO_NAME}")
         self.assertDictEqual(model.config, {"num": 10, "act": "gelu_fast"})
 
+    @retry_endpoint
     def test_push_to_hub(self):
         REPO_NAME = repo_name("PUSH_TO_HUB")
         model = DummyModel()
@@ -269,6 +274,7 @@ class HubKerasSequentialTest(HubMixingTestKeras):
         self.assertTrue(tf.reduce_all(tf.equal(new_model.weights[0], model.weights[0])))
         self.assertTrue(new_model.config == {"num": 10, "act": "gelu_fast"})
 
+    @retry_endpoint
     def test_push_to_hub(self):
         REPO_NAME = repo_name("PUSH_TO_HUB")
         model = self.model_init()
@@ -291,6 +297,7 @@ class HubKerasSequentialTest(HubMixingTestKeras):
 
         self._api.delete_repo(name=f"{REPO_NAME}", token=self._token)
 
+    @retry_endpoint
     def test_push_to_hub_tensorboard(self):
         os.makedirs(f"{WORKING_REPO_DIR}/tb_log_dir")
         with open(f"{WORKING_REPO_DIR}/tb_log_dir/tensorboard.txt", "w") as fp:
@@ -317,6 +324,7 @@ class HubKerasSequentialTest(HubMixingTestKeras):
 
         self._api.delete_repo(name=f"{REPO_NAME}", token=self._token)
 
+    @retry_endpoint
     def test_push_to_hub_model_kwargs(self):
         REPO_NAME = repo_name("PUSH_TO_HUB")
         model = self.model_init()
